requirements.txt
flask~=3.1.0
flask-cors
python-dotenv~=1.0.1
mysql-connector-python~=9.3.0
cryptography~=44.0.1
openai~=1.33.0
azure-cognitiveservices-speech
pydub~=0.25.1
requests~=2.31.0